Each line begins with a "strut" whose height determines the height of the line's bounding box. Previously, this strut included spaceabovelines, and it included abovelinestextheight if any, but it did not include abovelinestextraise. As a result, the user could sometimes increase abovelinestextheight but not see any effect. This commit includes abovelinestextraise.
Likewise, the strut's height already includes additional space due to additionaltopspacethreshold, but this commit also includes additional space due to additionaltopspacealtthreshold. The two increases aren't additive. The spaceabovelines is raised by the additional space due to additionaltopspacethreshold, and the abovelinestext is raised by additional space due to additionaltopspacealtthreshold, and whichever of the spaceabovelines and the top of the abovelinestext determines the top of the strut.