Fixes issue #1652, which was causing an error on scores with two line breaks in a row, (`(Z) (Z)`).
The root of the problem seems to be a bug in LuaTeX. The following MWE illustrates the bug:
```
\localrightbox{xyz}%
\hfill\break
\localleftbox{abc}%
\bye
```

This produces the errors
```
warning  (nodes): attempt to copy free hlist node 399, ignored

warning  (nodes): attempt to copy free hlist node 369, ignored
```
and the `\localleftbox` is not updated.

The bug was happening here when the `\localleftbox` was updated on an empty line. The simplest fix is just to not update the `\localleftbox` (it appears not to be needed after #1641).